Create feature to update contact by phone number

+++ src/main/java/com/api_contact_manager/models/Contact.java
@@ -13,7 +13,7 @@ public class Contact {
 
     @Id
     @Column(name="phone_number")
-    private int phoneNumber;
+    private Long phoneNumber;
     @Column(name="name")
     private String name;
 }
